Introduce 16-bit moves (#420)
bench: 5959962

+++ src/history.h
@@ -36,16 +36,16 @@ extern int16_t captureHistory[12][64][13];
 extern int16_t krpCorrhist[2][16384];
 
 int scaledBonus(int score, int bonus, int gravity);
-void updateQuietMoveHistory(int bestMove, int side, int depth, moves *badQuiets, board *pos);
-void updatePawnHistory(board *pos, int bestMove, int depth, moves *badQuiets);
-void updateSingleCHScore(board *pos, int move, const int offSet, const int bonus, int quiet_hist_score);
-int getAllCHScore(board *pos, int move, int quiet_hist_score);
-void updateAllCH(board *pos, int move, int bonus, int quiet_hist_score);
+void updateQuietMoveHistory(uint16_t bestMove, int side, int depth, moves *badQuiets, board *pos);
+void updatePawnHistory(board *pos, uint16_t bestMove, int depth, moves *badQuiets);
+void updateSingleCHScore(board *pos, uint16_t move, const int offSet, const int bonus, int quiet_hist_score);
+int getAllCHScore(board *pos, uint16_t move, int quiet_hist_score);
+void updateAllCH(board *pos, uint16_t move, int bonus, int quiet_hist_score);
 int getHistoryBonus(int depth);
-void updateContinuationHistory(board *pos, int bestMove, int depth, moves *badQuiets, int quiet_hist_score);
-int getContinuationHistoryScore(board *pos, int offSet, int move);
-void updateCaptureHistory(board *position, int bestMove, int depth);
-void updateCaptureHistoryMalus(board *position, int depth, moves *noisyMoves, int bestMove);
+void updateContinuationHistory(board *pos, uint16_t bestMove, int depth, moves *badQuiets, int quiet_hist_score);
+int getContinuationHistoryScore(board *pos, int offSet, uint16_t move);
+void updateCaptureHistory(board *position, uint16_t bestMove, int depth);
+void updateCaptureHistoryMalus(board *position, int depth, moves *noisyMoves, uint16_t bestMove);
 void clearQuietHistory(void);
